Fix 50013 missing permissions error for embed messages
If the bot was assigned a role that had no overrides for the embed permission set while the default role had it disabled or at the very least not enabled, the bot would throw an error when attempting to send embed messages, this should longer be the case as both the channel permissions for the bot, and the default "everyone" role is now checked at a raw permission level.

JDA should be checking this things internally but apprently it doesn't really work, or I'm doing it wrong, either way this works as a fix.