Create an undertale text box command
I would love to test it on my computer, however there seems to be no place to insert a Discord bot token at. It's supposed to return an image with some text on it by calling a website that uses javascript to build out an image. Potential issues could be separation of parameters, however it seems to be fine in the meme command. For now it only does vanilla undertale characters. If the provided character is not recognized, then no character image shows up in the text box.