Create undertale text box command (#40)
* Create an undertale text box command

I would love to test it on my computer, however there seems to be no place to insert a Discord bot token at. It's supposed to return an image with some text on it by calling a website that uses javascript to build out an image. Potential issues could be separation of parameters, however it seems to be fine in the meme command. For now it only does vanilla undertale characters. If the provided character is not recognized, then no character image shows up in the text box.

* Add .java extension

* Fix formatting and method placement

* Fix file being generated and sent correctly + Error messages

* Modify the UndertaleTextBoxCommand

Rather than listing out the characters, I decided to let a person who knows them rely on the API's list. If a person wants a custom character to be displayed instead, then they must provide a URL pointing to an image file of the character they want displayed. The list all valid character method exists so may be changed at a later date. I also tried my best to provide a description and give credit where credit is due.

* Fix character list working correctly + Sort characters alphabetically

* Do a bit of housekeeping to cleanup the code a bit

* Change description, usage, and example to fit more in an embed message

* Move the page under the list sub-prefix

* Move messages to the language files

* Move message query to the template url

* Bump version