Add a random comic argument to the XKCD command (#44)
* Add a random comic command to XKCD

The algorithm it uses is rather simple. If the argument provided to the command is not a number, then a random comic is retrieved. In order to find a random comic, it first finds the latest comic number. Next the java.util.Random class is utilized to generate a random number from 1 to the latest comic number, then a request is made with that number. Otherwise, if the argument is a number, then that comic number is retrieved.

* Add formatting to the code
* Add example usage and fix usage instructions
* Change random to use the random utility instead
* Add checks to see if the given argument is valid
* Bump version