Create a join date command
I am a little new to the OffsetDateTime class so hopefully I chose an appropriate formatter for it. I am have worked with the normal Date class previously. This command should help moderators that need to see how long a user has been a member of the server, in the event they use how long a user has been in the server to determine what role they should get.