Add full Heroku support for AvaIre (#53)
# Not ready for merge yet!

This pull request will finalize Avaire to be deployed on Heroku, using the app.json.

This pull also adds the wrapper.jar file. What is required for Heroku/windows/anyhost to compile AvaIre from source.

As of yet, i'm unable to get mysql working. I don't think it's smart to merge this unless this part works.

When AvaIre recieved this pull, i'll add additional documentation aswell as creating an issue on this GitHub explaining the little parts that aren't working as of yet. (The API for example, or the mysql string (from JawDB)).

### Please check this pull troughly!
###### ~~I'm still a noob with GitHub.~~