Fix some Heroku configuration (#55)
Added the last bits of the environment variables.
After this is pulled, i retest everything once again. And make an issue with all the things that are left to do for Heroku.
These, are not neceassry to run AvaIre on Heroku. They will just be an improvement as in:
Faster boot time
Being able to use the API
Being able to use plugins using env variables

As, Watchdog is simalair setup as AvaIre. I will soonish also make a pull for that repo aswell. Where you can build directly from source. Downloading AvaIre nightly build. For the people that don't want to source edit, but like the easy deployment (a best of both worlds?).