Changed web back to worker
I want to contain the easy use of Heroku.
Users who WANT to use the API, can turn worker off, and enable WEB worker instead.
Docs should suggest solutions how to keep this WEB worker up. My suggestions would be using https://kaffeine.herokuapp.com/ or Uptimerobot. First one is the easiest but not the most reliable in my eyes. But seems to be doing the job good enough. Second is my preferd option as it also mails the user optionally if the site goes down.