Add ability to repeat the currently playing song (#76)
This PR closes #46 and adds the ability to repeat the currently playing song on loop. In order to make this work, I had to rewrite RepeatMusicQueueCommand and change the queue to use a double ended queue. There's still some testing that needs to be done to make sure everything works and looks ok. I would also appreciate it if you would look over everything else so nothing is accidentally broken by this change.

TODO:

- [x] Add basic functionality
- [x] Add functionality to display current repeat mode
- [x] Rewrite usage instructions
- [x] Make better success info
- [x] Fix duplicate now playing messages
- [x] Test repeatsongs command in different scenarios
- [x] Make sure the change to deque doesn't break other commands and methods