Create pun command (#97)
* Create a user count command

It returns the number of users in the current guild.

* Create an advice command

It retrieves random advice from an Advice API.

* Revert "Create a user count command"

This reverts commit 8f9c517226fe629efbe9c9abb5f28e759937a1fe.

* Create a pun command

This command connects to the ICanHazDadJoke Api and can either retrieve a random pun or a list of puns based on the search criteria.

* Tidy up the commands

The Advice now doesn't have redundant info. The Pun command now uses a service now and sends a warning if the returned list is empty.

* Update PunCommand.java

Remove another character.

* Update PunCommand.java

The help command should work now and the paginator footer should be perfect finally.

* Advice command disabled to due NSFW

After looking at the api a lot more closely and getting embarrased while testing and reading the advice it returned, I decided to make it sure it only ran in NSFW channels.

* Update nl_NL.yml

Forgot to paste it in this file.

* Hopefully fix all remaining issues

As the advice command's ssl certificate was self-signed and overall problematic, it was disposed of. Camel-case lettering was used in the service request. Brackets were used to mark optional arguments. Lastly hopefully I fixed the Yaml file.

* Fix file formatting

* Rewrite pun command to return one result with query

The requester is also added to the embeds now.

* Add custom user agent according to the API regulations

* Remove new line added to the Chinese language file

* Delete unused travis file

Not sure why this was added to PR?

* Bump version