Create set role color command and color utility (#104)
* Prepare for a role color command

In addition the color util system was created.

* Create the set role color command

Fix the invalid color error message to be more clarifying.

* Fix formatting

* Fix bugs

First and foremost the permission required to run the command required the ability to manage roles. Then checking the permissions of both the bot and user permission were done. Lastly confirmation that the role color was successfully changed was added.

* Fix formatting + Cleanup some code

* Cleanup success and error messages

* Change manage role middleware check to include the bot

* Bump version