Resolve major issues
I had set the experience text color to the main color when it was initialized when I should have left it as null. I also had previously tped the wrong numbers in the YAML file. Lastly the files are now copied to the backgrounds folder and the file system ones are loaded first.