Improved script to the max
So the past hours i made the following happen:
we clone avaire to get the .git folder.
After compiling we delete this folder, as i rather delete and create a new one for the website, then merge and hope it's all clean.
We also delete everything that isn't hidden except for the file names after each ! (see line 2).
Then we clone the website repo, remove the procfile b/c it's wrong for this build. We move everything out of this folder and delete it.
We run both avaire to get the configs and to get the commandmap.json, we move that to the right place and it's all done.

What does this bring?
Shorter steps, flexible if another exception needs to be made. Just add after a ", this: ` ! -name "filenameorwhatever"` and it's done.
Config files are done automatically, so no maintance on that anymore!

Hope it's to your liking, it's been tested througly.