Move hidden file for Heroku buildpack (#118)
* Also move hidden files

This is the last pull to fix the small issues that have come with moving to the new deploy strategie.

I also found some improvements out, i list them down here, if you confirm it's true that avaire makes the missing .yml files herself.
If your ok to have the update.sh also on the website repo i can remove that command here aswell.

Proposed improvements:
i could run avaire a second time after the commandmap.json, so it makes the following automatically and more scableable:
  
 mv src/main/resources/config.yml ../config.yml
 mv src/main/resources/constants.yml ../constants.yml

  mv ../config.yml .
  mv ../constants.yml .

I just need to know for sure, avaire generates the missing yml files if it's ran withoud them?

mv update.sh ../update.sh

can maybe be removed if your ok to have this file in the website repo aswell!

* Error less moving files!

* Generate files with avaire

Instead of moving it ourselfs, we let avaire do it :D

* Optimized

Hope to find a way to exclude files from being deleted more efficiently. Heroku isn't working nicely with me -v-

* Improved script to the max

So the past hours i made the following happen:
we clone avaire to get the .git folder.
After compiling we delete this folder, as i rather delete and create a new one for the website, then merge and hope it's all clean.
We also delete everything that isn't hidden except for the file names after each ! (see line 2).
Then we clone the website repo, remove the procfile b/c it's wrong for this build. We move everything out of this folder and delete it.
We run both avaire to get the configs and to get the commandmap.json, we move that to the right place and it's all done.

What does this bring?
Shorter steps, flexible if another exception needs to be made. Just add after a ", this: ` ! -name "filenameorwhatever"` and it's done.
Config files are done automatically, so no maintance on that anymore!

Hope it's to your liking, it's been tested througly.

* Update buildpack-run.sh