Rewrite rank backgrounds handler (#108)
* Rewrite the background rank handler

No extra commits were provided in the making of this pull request. The background rank color class was made into a concrete class and an loader was unsuccessfully made.

* Make it actually compile

I forgot to check to see if I included all the files when I made my pull request. I forgot some apparently.

* Update DiscordDarkColors.yml

Actually make sure everything was the right type.

* Finally find all the files in the jar

Now all that's left is getting the classes and turning the content into YAML files then making sure they look right.

* Remove the link to the Spring Framework repo

It can easily be found on maven, a url is not needed.

* Fix the experience color

It looks like I was overwriting the experience foreground each time. Oops.

* Add Discord Light theme

* 8 Bit mountain colors

Create the eight bit colors and allow the folder location to only be changed in one spot.

* Finalize my changes

Everything looks ready.

* Fix my spelling

I had separator spelled wrong on every file.

* Load and render external files

There things should be able to loaded both inside of the AvaIre jar and external file system.

* Update RankBackgroundLoader.java

Restore the proper name to the variable.

* Update RankBackground.java

Actually read if it was loaded externally.

* Update RankBackgroundHandler.java

We don't need to know the name of the file that crashed.

* Push the background ranks into backgrounds

They are now in the right folder.

* Finally finished

I got a little ahead of myself and needed to revert the changes to the random dog command and make sure AvaIre actually loads the backgrounds. Oops.

* Resolve major issues

I had set the experience text color to the main color when it was initialized when I should have left it as null. I also had previously tped the wrong numbers in the YAML file. Lastly the files are now copied to the backgrounds folder and the file system ones are loaded first.

* Make RankHandler a singleton

It appears to officially be a singleton now.

* Remove the unnecessary lines

* Revert RandomDog command

It's back to a singleton list.

* Remove the default background

It has been moved to the resource folder with the rest of them.

* Remove the temporary variable.

* Make the code compile

I thought I fixed it last night.

* Update RankBackgroundHandler.java

Rather than create a new instance, I should have just looked up what it was returning in the first place.

* Fix code formatting

* Update RankBackgroundHandler.java

I think this should let users remove the default provided backgrounds now.

* Update RankBackgroundHandler.java

Fix the backgrounds from being automatically copied over if the folder already exists.

* Add backgrounds directory to the git ignore file

* Fix rank background color class comments and formatting

* Move rank background out of the contracts package

The class has been changed from an abstract class to a public class that is used and instantiated by other classes, so it's not a contract anymore.

* Refactor and fix code formatting in public imagegen classes

* Rename background file names

* Remove unnecessary full static call to itself in rank background colors

* Fix git ignore file to only work on the top most level

* Make RankBackgroundHandler a proper singleton

Only the instance field is static and the light theme gained the experienceTextColor heading.

* Actually use the Used ids field

I wondered why that didn't actually work.  Guess that field just needed used.

* Fixed the documentation in a few places

This should be a little better now.

* Improve Javadocs a bit

More detail was put into the names of the variables as well as some of the descriptions.

* Properly explain where it was being loaded from

I was outputting it was being loaded from the resource folder when it was actually being loaded from file system.

* Fix backgrounds having the same name

The namesToCost variable wasn't being updated until after all the files were on a queue to be added and sorted.

* Add fallback to embedded message if background render errors out

With the new system images can be changed on the fly, so in the event that a background has been registered with an image resource that doesn't actually exists, this will just fall back to the embed version of the command, instead of showing nothing to the user.

* Fix code formatting

* Remove unused code + Cleanup some temporary variables

* Fix documentation to provide more useful information

* Add an indexer system

A file get written to every time backgrounds are copied over. If it is not on the list, they get copied over.

* Removed the extra line character

It wasn't actually necessary to have an extra new line when you go to append to it.

* Remove extra new lines

Those two new lines shouldn't be there.

* Use a yaml file instead

The index has also been moved up a file and renamed to background-index.

* Create additional documentation

It should be a little better now.

* Add additional documentation

It should be more thorough and possibly better, even for private methods.

* Remove unused methods

I removed all the unused methods specifically made when trying to brainstorm what the RankBackgroundLoader should use.

* Actually use a logger

It was easier while writing this to print directly to the console.

* Load conditionally

If the background folder didn't exist at boot time, then it would only load the backgrounds from the jar.

* Use the yaml configuration properly

Everything was under this one heading and I didn't think to include the top-most heading. Using get rather contains seems less problematic as I would have to be aware of the separator.  I also renamed background-index to index.

* Change the block comment

It needed to say what actually occured.

* Use logger on RankCommand

Hopefully this provides good logged error messages.

* Use the actual path separator

I can't believe I didn't actually notice this before.

* Fix formatting and codedoc

* Replace get with contains

It seriously should not have been as simple as using contains as get.

* It was a list the whole time

The way it was being set was a list and suprisingly I didn't even notice that was available the whole time. I set the values in a list to the index object when I wrote to it; of course it had to be read as a list.

* Refactor rank background handler to remove unnecessary code

* Fix formatting and logging output in rank background handler

* Move background index into the backgrounds directory

* Change invalid config log output to warn to make it more visible

* Refactor image generation classes and rank command

This changes how the comments are order to make them look a bit nicer to the eyes, as-well as getting rid of unused code, and removing some unnecessary code.

* Add space between static and non-static properties

* Create more clear failed log messages for invalid rank background files

* Bump version