Changed jawsdb to the maria version
Java related changes to no longer look for the jawsdb env variable
but instead listen for the maria db version.

This will make auto-updateable avaire's on Heroku no longer work, but if they add the maria db from jawsdb, it will work. They have to somehow migrate the db content themselfs... Or decided not to update.