Fix for meme generator command
memegen.link API generates links in the format of "http://" instead of "https://" that this code is considering. This results in the loss of the first character in every single meme name, since it's directly considering the length of the "https" apiEndpoint string. This fix adds a check to use the correct length depending on whether the start of the links are http or https, so that the code will not break even when the API switches back to https-based links.